cpp impl: Split Code-Files
Split up files for a better code overview
