+++ src/cpp/src/internal/kvs_helper.hpp
@@ -15,6 +15,9 @@
 
 #include <sstream>
 #include <string>
+#include "error.hpp"
+#include "kvsvalue.hpp"
+#include "score/json/json_parser.h" /* For JSON Any Type */
 
 /* 
  * This header defines helper functions used internally by the Key-Value Store (KVS) implementation.
@@ -33,6 +36,7 @@ uint32_t parse_hash_adler32(std::istream& in);
 uint32_t calculate_hash_adler32(const std::string& data);
 std::array<uint8_t,4> get_hash_bytes_adler32(uint32_t hash);
 std::array<uint8_t,4> get_hash_bytes(const std::string& data);
+bool check_hash(const std::string& data_calculate, std::istream& data_parse);
 score::Result<KvsValue> any_to_kvsvalue(const score::json::Any& any);
 score::Result<score::json::Any> kvsvalue_to_any(const KvsValue& kv);
