+++ src/cpp/src/kvs.cpp
@@ -10,13 +10,11 @@
 *
 * SPDX-License-Identifier: Apache-2.0
 ********************************************************************************/
-
 #include <fstream>
 #include <iostream>
 #include <sstream>
-#include "kvs.hpp"
 #include "internal/kvs_helper.hpp"
-
+#include "kvs.hpp"
 
 //TODO Default Value Handling TBD
 //TODO Add Score Logging
@@ -33,434 +31,6 @@ namespace pers
 {
 namespace kvs 
 {
-
-/*********************** Hash Functions *********************/
-/*Adler 32 checksum algorithm*/ 
-// Optimized version: processes data in blocks to reduce modulo operations
-uint32_t calculate_hash_adler32(const string& data) {
-    constexpr size_t ADLER32_NMAX = 5552;
-    constexpr uint32_t ADLER32_BASE = 65521;
-    uint32_t a = 1, b = 0;
-    size_t len = data.size();
-    size_t i = 0;
-
-    // Process in blocks of 5552 bytes (as recommended for Adler-32)
-    while (len > 0) {
-        size_t tlen = len > ADLER32_NMAX ? ADLER32_NMAX : len;
-        len -= tlen;
-        for (size_t j = 0; j < tlen; ++j, ++i) {
-            a += static_cast<unsigned char>(data[i]);
-            b += a;
-        }
-        a %= ADLER32_BASE;
-        b %= ADLER32_BASE;
-    }
-    return (b << 16) | a;
-}
-
-/*Parse Adler32 checksum Byte-Array to uint32 */ 
-uint32_t parse_hash_adler32(std::istream& in)
-{
-    std::array<uint8_t,4> buf{};
-    in.read(reinterpret_cast<char*>(buf.data()), buf.size());
-
-    uint32_t value = (uint32_t(buf[0]) << 24)
-                   | (uint32_t(buf[1]) << 16)
-                   | (uint32_t(buf[2]) <<  8)
-                   |  uint32_t(buf[3]);
-    return value;
-}
-
-/* Split uint32 checksum in bytes for writing*/ 
-std::array<uint8_t,4> get_hash_bytes_adler32(uint32_t hash)
-{
-    std::array<uint8_t, 4> value = {
-        uint8_t((hash >> 24) & 0xFF),
-        uint8_t((hash >> 16) & 0xFF),
-        uint8_t((hash >>  8) & 0xFF),
-        uint8_t((hash      ) & 0xFF)
-    };
-    return value;
-}
-
-/***** Wrapper Functions for Hash *****/
-/* Wrapper Functions should isolate the Hash Algorithm, so that the algorithm can be easier replaced*/
-
-/* Wrapper Function to get checksum in bytes*/
-std::array<uint8_t,4> get_hash_bytes(const string& data)
-{
-    uint32_t hash = calculate_hash_adler32(data);
-    std::array<uint8_t, 4> value = get_hash_bytes_adler32(hash);
-    return value;
-}
-
-/* Wrapper Function to check, if Hash is valid*/
-bool check_hash(const string& data_calculate, std::istream& data_parse){
-    bool result;
-    uint32_t calculated_hash = calculate_hash_adler32(data_calculate);
-    uint32_t parsed_hash = parse_hash_adler32(data_parse);
-    if(calculated_hash == parsed_hash){
-        result = true;
-    }else{
-        result = false;
-    }
-
-    return result; 
-}
-
-/*********************** Standalone Helper Functions *********************/
-
-/* Helper Function for Any -> KVSValue conversion */
-score::Result<KvsValue> any_to_kvsvalue(const score::json::Any& any){
-    score::Result<KvsValue> result = score::MakeUnexpected(ErrorCode::UnmappedError);
-    if (auto o = any.As<score::json::Object>(); o.has_value()) {
-        const auto& objAny = o.value().get();
-        auto type  = objAny.find("t");
-        auto value = objAny.find("v");
-        if (type != objAny.end() && value != objAny.end()) {
-            if (auto typeStr = type->second.As<std::string>(); typeStr.has_value()) {
-                const std::string_view typeStrV = typeStr.value().get();
-                const score::json::Any& valueAny = value->second;
-
-                if (typeStrV == "i32") {
-                    if (auto n = valueAny.As<int32_t>(); n.has_value()){
-                        result = KvsValue(static_cast<int32_t>(n.value()));
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "u32") {
-                    if (auto n = valueAny.As<uint32_t>(); n.has_value()) {
-                        result = KvsValue(static_cast<uint32_t>(n.value()));
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "i64") {
-                    if (auto n = valueAny.As<int64_t>(); n.has_value()) {
-                        result = KvsValue(static_cast<int64_t>(n.value()));
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "u64") {
-                    if (auto n = valueAny.As<uint64_t>(); n.has_value()) {
-                        result = KvsValue(static_cast<uint64_t>(n.value()));
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "f64") {
-                    if (auto n = valueAny.As<double>(); n.has_value()) {
-                        result = KvsValue(n.value());
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "bool") {
-                    if (auto b = valueAny.As<bool>(); b.has_value()) {
-                        result = KvsValue(b.value());
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "str") {
-                    if (auto s = valueAny.As<std::string>(); s.has_value()) {
-                        result = KvsValue(s.value().get());
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "null") {
-                    if (valueAny.As<score::json::Null>().has_value()) {
-                        result = KvsValue(nullptr);
-                    }
-                    else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "arr") {
-                    if (auto l = valueAny.As<score::json::List>(); l.has_value()) {
-                        KvsValue::Array arr;
-                        bool error = false;
-                        for (auto const& elem : l.value().get()) {
-                            auto conv = any_to_kvsvalue(elem);
-                            if (!conv) {
-                                error = true;
-                                result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                                break;
-                            }
-                            arr.emplace_back(std::move(conv.value()));
-                        }
-                        if (!error){
-                            result = KvsValue(std::move(arr));
-                        }
-                    } else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                }
-                else if (typeStrV == "obj") {
-                    if (auto obj = valueAny.As<score::json::Object>(); obj.has_value()) {
-                        KvsValue::Object map;
-                        bool error = false;
-                        for (auto const& [key, valAny] : obj.value().get()) {
-                            auto conv = any_to_kvsvalue(valAny);
-                            if (!conv) {
-                                error = true;
-                                result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                                break;
-                            }
-                            map.emplace(key.GetAsStringView().to_string(), std::move(conv.value()));
-                        }
-                        if (!error) {
-                            result = KvsValue(std::move(map));
-                        }
-                    } else {
-                        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    }
-                } else {
-                    result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                }
-            } else {
-                result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-            }
-        } else {
-            result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-        }
-    } else {
-        result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-    }
-
-    return result;
-}
-
-/* Helper Function for KVSValue -> Any conversion */
-score::Result<score::json::Any> kvsvalue_to_any(const KvsValue& kv) {
-    score::Result<score::json::Any> result = score::MakeUnexpected(ErrorCode::UnmappedError);
-    bool error = false;
-    score::json::Object obj;
-    switch (kv.getType()) {
-        case KvsValue::Type::i32: {
-            obj.emplace("t", score::json::Any(std::string("i32")));
-            obj.emplace("v", score::json::Any(static_cast<int32_t>(std::get<int32_t>(kv.getValue()))));
-            break;
-        }
-        case KvsValue::Type::u32: {
-            obj.emplace("t", score::json::Any(std::string("u32")));
-            obj.emplace("v", score::json::Any(static_cast<uint32_t>(std::get<uint32_t>(kv.getValue()))));
-            break;
-        }
-        case KvsValue::Type::i64: {
-            obj.emplace("t", score::json::Any(std::string("i64")));
-            obj.emplace("v", score::json::Any(static_cast<int64_t>(std::get<int64_t>(kv.getValue()))));
-            break;
-        }
-        case KvsValue::Type::u64: {
-            obj.emplace("t", score::json::Any(std::string("u64")));
-            obj.emplace("v", score::json::Any(static_cast<uint64_t>(std::get<uint64_t>(kv.getValue()))));
-            break;
-        }
-        case KvsValue::Type::f64: {
-            obj.emplace("t", score::json::Any(std::string("f64")));
-            obj.emplace("v", score::json::Any(std::get<double>(kv.getValue())));
-            break;
-        }
-        case KvsValue::Type::Boolean: {
-            obj.emplace("t", score::json::Any(std::string("bool")));
-            obj.emplace("v", score::json::Any(std::get<bool>(kv.getValue())));
-            break;
-        }
-        case KvsValue::Type::String: {
-            obj.emplace("t", score::json::Any(std::string("str")));
-            obj.emplace("v", score::json::Any(std::get<std::string>(kv.getValue())));
-            break;
-        }
-        case KvsValue::Type::Null: {
-            obj.emplace("t", score::json::Any(std::string("null")));
-            obj.emplace("v", score::json::Any(score::json::Null{}));
-            break;
-        }
-        case KvsValue::Type::Array: {
-            obj.emplace("t", score::json::Any(std::string("arr")));
-            score::json::List list;
-            for (auto& elem : std::get<KvsValue::Array>(kv.getValue())) {
-                auto conv = kvsvalue_to_any(elem);
-                if (!conv) {
-                    result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    error = true;
-                    break;
-                }
-                list.push_back(std::move(conv.value()));
-            }
-            if (!error) {
-                obj.emplace("v", score::json::Any(std::move(list)));
-            }
-            break;
-        }
-        case KvsValue::Type::Object: {
-            obj.emplace("t", score::json::Any(std::string("obj")));
-            score::json::Object inner_obj;
-            for (auto& [key, value] : std::get<KvsValue::Object>(kv.getValue())) {
-                auto conv = kvsvalue_to_any(value);
-                if (!conv) {
-                    result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-                    error = true;
-                    break;
-                }
-                inner_obj.emplace(key, std::move(conv.value()));
-            }
-            if (!error) {
-                obj.emplace("v", score::json::Any(std::move(inner_obj)));
-            }
-            break;
-        }
-        default: {
-            result = score::MakeUnexpected(ErrorCode::InvalidValueType);
-            error = true;
-            break;
-        }
-    }
-
-    if (!error) {
-        result = score::json::Any(std::move(obj));
-    }
-
-    return result;
-}
-
-
-/*********************** Error Implementation *********************/
-std::string_view MyErrorDomain::MessageFor(score::result::ErrorCode const& code) const noexcept
-{
-    std::string_view msg;
-    switch (static_cast<ErrorCode>(code))
-    {
-        case ErrorCode::UnmappedError:
-            msg = "Error that was not yet mapped";
-            break;
-        case ErrorCode::FileNotFound:
-            msg = "File not found";
-            break;
-        case ErrorCode::KvsFileReadError:
-            msg = "KVS file read error";
-            break;
-        case ErrorCode::KvsHashFileReadError:
-            msg = "KVS hash file read error";
-            break;
-        case ErrorCode::JsonParserError:
-            msg = "JSON parser error";
-            break;
-        case ErrorCode::JsonGeneratorError:
-            msg = "JSON generator error";
-            break;
-        case ErrorCode::PhysicalStorageFailure:
-            msg = "Physical storage failure";
-            break;
-        case ErrorCode::IntegrityCorrupted:
-            msg = "Integrity corrupted";
-            break;
-        case ErrorCode::ValidationFailed:
-            msg = "Validation failed";
-            break;
-        case ErrorCode::EncryptionFailed:
-            msg = "Encryption failed";
-            break;
-        case ErrorCode::ResourceBusy:
-            msg = "Resource is busy";
-            break;
-        case ErrorCode::OutOfStorageSpace:
-            msg = "Out of storage space";
-            break;
-        case ErrorCode::QuotaExceeded:
-            msg = "Quota exceeded";
-            break;
-        case ErrorCode::AuthenticationFailed:
-            msg = "Authentication failed";
-            break;
-        case ErrorCode::KeyNotFound:
-            msg = "Key not found";
-            break;
-        case ErrorCode::KeyDefaultNotFound:
-            msg = "Key default value not found";
-            break;
-        case ErrorCode::SerializationFailed:
-            msg = "Serialization failed";
-            break;
-        case ErrorCode::InvalidSnapshotId:
-            msg = "Invalid snapshot ID";
-            break;
-        case ErrorCode::ConversionFailed:
-            msg = "Conversion failed";
-            break;
-        case ErrorCode::MutexLockFailed:
-            msg = "Mutex failed";
-            break;
-        case ErrorCode::InvalidValueType:
-            msg = "Invalid value type";
-            break;
-        default:
-            msg = "Unknown Error!";
-            break;
-    }
-
-    return msg;
-}
-
-score::result::Error MakeError(ErrorCode code, std::string_view user_message) noexcept
-{
-    return {static_cast<score::result::ErrorCode>(code), my_error_domain, user_message};
-}
-
-
-/*********************** KVS Builder Implementation *********************/
-KvsBuilder::KvsBuilder(const InstanceId& instance_id)
-    : instance_id(instance_id)
-    , need_defaults(false)
-    , need_kvs(false)
-    , directory("./data_folder/") /* Default Directory */
-{}
-
-KvsBuilder& KvsBuilder::need_defaults_flag(bool flag) {
-    need_defaults = flag;
-    return *this;
-}
-
-KvsBuilder& KvsBuilder::need_kvs_flag(bool flag) {
-    need_kvs = flag;
-    return *this;
-}
-
-KvsBuilder& KvsBuilder::dir(std::string&& dir_path) {
-    this->directory = std::move(dir_path);
-    return *this;
-}
-
-
-score::Result<Kvs> KvsBuilder::build() {
-    score::Result<Kvs> result = score::MakeUnexpected(ErrorCode::UnmappedError);
-
-    /* Use current directory if empty */
-    if ("" == directory) {
-        directory = "./";
-    }
-    
-    result = Kvs::open(
-        instance_id,
-        need_defaults ? OpenNeedDefaults::Required : OpenNeedDefaults::Optional,
-        need_kvs      ? OpenNeedKvs::Required      : OpenNeedKvs::Optional,
-        std::move(directory)
-    );
-    
-    return result;
-}
-
 
 /*********************** KVS Implementation *********************/
 Kvs::~Kvs(){
@@ -569,7 +139,7 @@ score::Result<std::unordered_map<string, KvsValue>> Kvs::open_json(const score::
 {   
     score::filesystem::Path json_file = prefix.Native() + ".json";
     score::filesystem::Path hash_file = prefix.Native() + ".hash";
-    string data;
+    std::string data;
     bool error = false; /* Error flag */
     bool new_kvs = false; /* Flag to check if new KVS file is created*/
     score::Result<std::unordered_map<string, KvsValue>> result = score::MakeUnexpected(ErrorCode::UnmappedError);
@@ -704,7 +274,7 @@ score::Result<std::vector<std::string_view>> Kvs::get_all_keys() {
 }
 
 /* Check if a key exists*/
-score::Result<bool> Kvs::key_exists(const string_view key) {
+score::Result<bool> Kvs::key_exists(const std::string_view key) {
     score::Result<bool> result = score::MakeUnexpected(ErrorCode::UnmappedError);
     std::unique_lock<std::mutex> lock(kvs_mutex, std::try_to_lock);
     if (lock.owns_lock()) {
@@ -815,7 +385,7 @@ score::ResultBlank Kvs::set_value(const std::string_view key, const KvsValue& va
 }
 
 /* Remove a key-value pair*/
-score::ResultBlank Kvs::remove_key(const string_view key) {
+score::ResultBlank Kvs::remove_key(const std::string_view key) {
     score::ResultBlank result = score::MakeUnexpected(ErrorCode::UnmappedError);
     std::unique_lock<std::mutex> lock(kvs_mutex, std::try_to_lock);
     if (lock.owns_lock()) {
