+++ src/cpp/tests/bm_kvs.cpp
@@ -16,10 +16,11 @@
 
 #define private public
 #define final 
-#include "kvs.hpp"
+#include "kvsbuilder.hpp"
 #undef private
 #undef final
 #include "internal/kvs_helper.hpp"
+using namespace score::mw::pers::kvs;
 
 static void BM_get_hash_bytes(benchmark::State& state) {
     // Prepare a test string of configurable size
