+++ src/cpp/tests/test_kvs.cpp
@@ -20,10 +20,10 @@
 #include <string>
 #include <unistd.h>
 
-/* Change Private Members and final to public to allow access to member variables and derive from kvsvalue in unittests*/
+/* Change Private Members and final to public to allow access to member variables (kvs and kvsbuilder) and derive from kvsvalue in unittests*/
 #define private public
 #define final 
-#include "kvs.hpp"
+#include "kvsbuilder.hpp"
 #undef private
 #undef final
 #include "internal/kvs_helper.hpp"
